+++ src/main/java/com/icthh/xm/uaa/service/permission/PermissionUpdateService.java
@@ -6,12 +6,12 @@ import com.icthh.xm.commons.security.XmAuthenticationContextHolder;
 import com.icthh.xm.commons.tenant.TenantContextHolder;
 import com.icthh.xm.commons.tenant.TenantContextUtils;
 import com.icthh.xm.lep.api.LepManager;
-import com.icthh.xm.uaa.service.*;
+import com.icthh.xm.uaa.service.AuthenticationService;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.context.SecurityContextImpl;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 
 import java.util.Collection;
 import java.util.Set;
@@ -24,7 +24,7 @@ import static com.icthh.xm.commons.lep.XmLepConstants.THREAD_CONTEXT_KEY_TENANT_
 /**
  * //todo V: doc
  */
-@Component
+@Service
 @RequiredArgsConstructor
 public class PermissionUpdateService {
