+++ src/main/java/com/icthh/xm/uaa/service/permission/PermissionsConfigMode.java
@@ -1,4 +1,4 @@
-package com.icthh.xm.uaa.service;
+package com.icthh.xm.uaa.service.permission;
 
 /**
  * Describes configuration mode
