+++ src/main/java/com/icthh/xm/uaa/service/permission/PermissionsConfigurationProvider.java
@@ -1,4 +1,4 @@
-package com.icthh.xm.uaa.service;
+package com.icthh.xm.uaa.service.permission;
 
 import com.icthh.xm.commons.permission.domain.Permission;
 import com.icthh.xm.commons.permission.domain.Role;
