+++ src/test/java/com/icthh/xm/uaa/service/permission/PermissionUpdateServiceIntTest.java
@@ -6,7 +6,6 @@ import com.icthh.xm.commons.permission.domain.Privilege;
 import com.icthh.xm.commons.tenant.TenantContextHolder;
 import com.icthh.xm.uaa.UaaApp;
 import com.icthh.xm.uaa.config.xm.XmOverrideConfiguration;
-import com.icthh.xm.uaa.service.*;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
