+++ src/test/java/com/icthh/xm/uaa/web/rest/RoleResourceIntTest.java
@@ -5,8 +5,8 @@ import com.icthh.xm.commons.config.client.repository.TenantConfigRepository;
 import com.icthh.xm.commons.i18n.error.web.ExceptionTranslator;
 import com.icthh.xm.uaa.UaaApp;
 import com.icthh.xm.uaa.config.xm.XmOverrideConfiguration;
-import com.icthh.xm.uaa.service.PermissionsConfigMode;
-import com.icthh.xm.uaa.service.PermissionsConfigModeProvider;
+import com.icthh.xm.uaa.service.permission.PermissionsConfigMode;
+import com.icthh.xm.uaa.service.permission.PermissionsConfigModeProvider;
 import lombok.SneakyThrows;
 import org.junit.Before;
 import org.junit.Test;
